+++ features/support/env.rb
@@ -1,0 +1,2 @@
+$LOAD_PATH.unshift File.expand_path('lib', Dir.pwd)
+require 'cucumber_priority'
