+++ features/test.feature
@@ -1,0 +1,7 @@
+Feature: Serve coffee
+  Coffee should not be served until paid for
+  Coffee should not be served until the button has been pressed
+  If there is no coffee left then money should be refunded
+
+Scenario: Buy last coffee
+  Given there is a movie with a funny tone
