+++ lib/cucumber_priority.rb
@@ -4,5 +4,5 @@ require 'cucumber/rb_support/rb_language'
 require 'cucumber_priority/util'
 require 'cucumber_priority/ambiguous_error_ext'
 require 'cucumber_priority/rb_step_definition_ext'
-require 'cucumber_priority/support_code_ext'
+require 'cucumber_priority/rb_language'
